pyproject.toml
[project]
name = "iot-platform-demo"
version = "0.0.1"
description = "An IoT system demo"
authors = [
    {name = "nifrali", email = "francisco.liwa@gmail.com"},
]
dependencies = [
    "duckdb>=0.9.2",
    "httpx>=0.25.0",
    "prefect>=2.13.4",
    "python-dotenv>=1.0.0",
    "pandas==2.0.0",
    "fastapi>=0.103.2",
    "uvicorn>=0.23.2",
    "streamlit>=1.27.2",
    "plost>=0.2.5",
    "streamlit-folium>=0.15.0",
    "plotly>=5.17.0",
    "streamlit-plotly-events>=0.0.6",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

